requirements/dev.txt
-r base.txt

# django debug
django-debug-toolbar==1.3.0

# code-style
pep8==1.6.2

# interactive
ipython
